Fix bug in PKCS #7 padding implementation
Previously, if the input length was already a multiple of the block
size, no padding was added. Such an input with the last byte equal to
0x01 could not be stripped of padding unambiguously, since such a
trailing byte could also be interpreted as padding.

Now, if the input length is a multiple of the block size, a whole block
of padding is added.